+++ Kernel/System/Daemon/SchedulerDB.pm
@@ -1399,6 +1399,8 @@ sub CronTaskToExecute {
     # get needed objects
     my $CronEventObject = $Kernel::OM->Get('Kernel::System::CronEvent');
 
+    my %UsedTaskNames;
+
     CRONJOBKEY:
     for my $CronjobKey ( sort keys %{$Config} ) {
 
@@ -1409,20 +1411,22 @@ sub CronTaskToExecute {
 
         next CRONJOBKEY if !IsHashRefWithData($JobConfig);
 
-        if ( !$JobConfig->{Module} ) {
-            $Kernel::OM->Get('Kernel::System::Log')->Log(
-                Priority => 'error',
-                Message  => "Config option Daemon::SchedulerCronTaskManager::Task###$CronjobKey is invalid."
-                    . " Need 'Module' parameter!",
-            );
-            next CRONJOBKEY;
+        for my $Needed (qw(Module TaskName Function)) {
+            if ( !$JobConfig->{$Needed} ) {
+                $Kernel::OM->Get('Kernel::System::Log')->Log(
+                    Priority => 'error',
+                    Message  => "Config option Daemon::SchedulerCronTaskManager::Task###$CronjobKey is invalid."
+                        . " Need '$Needed' parameter!",
+                );
+                next CRONJOBKEY;
+            }
         }
 
-        if ( $JobConfig->{Module} && !$JobConfig->{Function} ) {
+        if ( $UsedTaskNames{ $JobConfig->{TaskName} } ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
                 Priority => 'error',
                 Message  => "Config option Daemon::SchedulerCronTaskManager::Task###$CronjobKey is invalid."
-                    . " Need 'Function' parameter!",
+                    . " TaskName parameter '$JobConfig->{TaskName}' is already used by another task!",
             );
             next CRONJOBKEY;
         }
@@ -1448,6 +1452,7 @@ sub CronTaskToExecute {
                 Params   => $JobConfig->{Params}   || '',
             },
         );
+        $UsedTaskNames{ $JobConfig->{TaskName} } = 1;
     }
 
     return 1;
@@ -2015,7 +2020,6 @@ sub RecurrentTaskExecute {
         }
     }
 
-    # get cache object
     my $CacheObject = $Kernel::OM->Get('Kernel::System::Cache');
 
     my $CacheKey = "$Param{TaskName}::$Param{TaskType}";